A bit of security

+++ generator.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 
 import os
+import sys
 import shutil
 import subprocess
 
@@ -25,8 +26,9 @@ class GitCommand(Command):
 
 
 class GitProject(object):
-    def __init__(self, folder_name):
+    def __init__(self, folder_name, always_proceed=False):
         self.folder_name = folder_name
+        self.always_proceed = always_proceed
         self.commands = [GitCommand('init')]
 
     def append_command(self, executable, *args):
@@ -41,21 +43,30 @@ class GitProject(object):
         self.commands.append(GitCommand('commit', '-m', file_root_name))
 
     def reset(self):
-        shutil.rmtree(self.folder_name, ignore_errors=True)
+        if not self.always_proceed:
+            answer = input(' ** WARNING : I am about to erase folder : {0}. Proceed ? [yN]'.format(self.folder_name))
+        else:
+            answer = 'y'
+        if answer == 'y':
+            shutil.rmtree(self.folder_name, ignore_errors=True)
+        else:
+            print(' Ok, fine. Not touching anything.')
+        return (answer == 'y')
 
     def initialize(self):
         os.mkdir(self.folder_name)
         os.chdir(self.folder_name)
 
     def create(self):
-        self.reset()
-        self.initialize()
-        for command in self.commands:
-            command.execute()
+        if self.reset():
+            self.initialize()
+            for command in self.commands:
+                command.execute()
 
 
 if __name__ == '__main__':
-    gp = GitProject('git-book-v2-chapter-3.1-branches-in-a-nutshell')
+    ap = True if len(sys.argv) > 1 and sys.argv[1] == '--always_proceed' else False
+    gp = GitProject('git-book-v2-chapter-3.1-branches-in-a-nutshell', always_proceed=ap)
     gp.append_new_committed_file_command('98ca9')
     gp.append_new_committed_file_command('34ac2')
     gp.append_new_committed_file_command('f30ab')
